requirements.txt
numexpr
pyTelegramBotAPI
python-decouple
